[libxl] Reserve memory for system usage
There is a small range of memory values that cause the guest
to crash during startup. It is possible for the libxl process
to fail to calculate free space requirements for internal
structures and allocations.

When checking if there is enough memory to support a guest, reserve
an extra 256MB for allocations and other necessary buffers.

Signed-off-by: Tyler McGavran <mcgavrant@ainfosec.com>