[key-functions] Handle case where dialog errors
If something trips ML, a user is prompted for a password and escape is
pressed, a case in key-functions prompt_recovery_password is hit that is
improperly handled, resulting in stdin being sent to the TTY since
openssl and cryptsetup are attempting to prompt for the password over
stdout, but is masked. Instead, check the RC value and handle the case
where dialog fails or escape is pressed.

Signed-off-by: esengardb <esengardb@ainfosec.com>